backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
overpy
osmnx
sentence-transformers
pinecone
langchain
langchain-groq
fpdf2



